Create augmentTest
This file is a test for augmenting our TRAINING data. It should never be run on our test data! In order to view the augmented images properly, rescaling has been commented out. When this code is used in the full scale run, it should first - uncomment scaling - and second - loop through all ten genre folders in the training directory, augmenting EVERY image there.